refactor(cloud-server): split the CloudSessionService god class
Behavior-preserving decomposition of the 2,718-line session-service.ts
(the audit-flagged "unreviewably large unit") into cohesive modules,
following the existing services/ pattern. Public API is byte-for-byte
unchanged — no caller or test edits.

- session-service-types.ts — shared types + promptParts/importAuditActor
  helpers + consts (re-exported for surface compatibility)
- services/principal-service.ts — ensurePrincipal + principal bootstrap/authz
  (owns the per-instance bootstrappedPrincipals TTL cache)
- services/session-import-service.ts — createImportedSession / completeSessionImport
  / recordImportFailed
- session-coordination-dispatch.ts — coordination-watch fan-out
- session-execution-operations.ts — command enqueue/execute + runtime-event engine
  (calls back into the facade's appendProjectedEvent)

session-service.ts: 2,718 -> 1,598 lines. The facade keeps thin delegators
with identical signatures; extracted files respect the 450-line/file modularity
budget (execution engine lives in src/ per the session-*-operations.ts pattern).

Gates: typecheck, eslint --max-warnings 0, knip, node (2238) + workspace (313)
suites, coverage inventory (92.0% >= 90%) all pass. 139 public members verified
identical vs base.
